refactor: 테스트 REST 관련 로직 클래스 분리
